refactor: apply to openapi spec
